sendsms.php: fix bug of wrong authentication into ini file

+++ script/sendsms.php
@@ -50,10 +50,14 @@ if ($handle = opendir(SMS_DIR))
 	$sender = $tmp[0];
 	// the prefix is managed by proxycti server. It is present in the file name (ex. 271-00393331234567)
 	$destination = $tmp[1];
-	send_sms($destination,$xbody,$xhost,$xusername,$xpassword,$lock);
-	unlink(SMS_DIR.'/'.$file);
-	$xbody = mysql_real_escape_string($xbody);
-	mysql_query("INSERT INTO sms_history (sender,destination,text,date,status) VALUES ('$sender','$destination','$xbody',now(),1)");
+	$res = send_sms($destination,$xbody,$xhost,$xusername,$xpassword,$lock);
+	if($res!=''){
+		unlink(SMS_DIR.'/'.$file);
+		$xbody = mysql_real_escape_string($xbody);
+		mysql_query("INSERT INTO sms_history (sender,destination,text,date,status) VALUES ('$sender','$destination','$xbody',now(),1)");
+	} else {
+		echo "sms send failed: check authentication\n";
+	}
 	sleep(5);
     }
 }
